backend/requirements.txt
kafka-python==2.0.2
python-dotenv==1.0.1
clickhouse-connect==0.8.14
websockets==15.0.1
requests==2.32.3
python-binance==1.0.28
pandas==2.2.3
Flask==3.1.0
clickhouse-driver==0.2.9
bcrypt==4.3.0
flask-cors==5.0.1
simplejson==3.20.1
